requirements.txt
streamlit
qrcode
io
cryptography
mysql-connector-python
pyzbar
pillow
random
pandas